requirements.txt
Flask==2.3.2
numpy
opencv-python
Pillow
scikit-learn
scikit-image
matplotlib
Werkzeug
gunicorn==25.1.1